+++ src/services/telemetry/frame_decoders/TeleMiniV3FrameDecoder.cpp
@@ -5,7 +5,7 @@ FlightSample TeleMiniV3FrameDecoder::decode(const Frame &frame) {
     sample.timestamp = get_numerical_field_le<int>(frame, 2, 2);
 
     sample.batteryVoltage = get_numerical_field_le<int>(frame, 6, 2);
-    sample.pressure = get_numerical_field_le<int>(frame, 12, 4) / 10;
+    sample.pressure = static_cast<double>(get_numerical_field_le<int>(frame, 12, 4)) / 10;
     sample.temperature = static_cast<double>(get_numerical_field_le<int16_t>(frame, 16, 2)) / 100;
     sample.altitude = get_numerical_field_le<int16_t>(frame, 22, 2);
